+++ mockserver-core/src/main/java/org/mockserver/collections/SubSetMatcher.java
@@ -1,10 +1,13 @@
 package org.mockserver.collections;
 
 import org.mockserver.matchers.RegexStringMatcher;
+import org.mockserver.model.NottableString;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+
+import static org.mockserver.model.NottableString.string;
 
 public class SubSetMatcher {
 
@@ -12,9 +15,10 @@ public class SubSetMatcher {
         boolean result = true;
         Set<Integer> matchingIndexes = new HashSet<>();
         for (ImmutableEntry subsetItem : subset) {
-            Set<Integer> subsetItemMatchingIndexes = matchesIndexes(regexStringMatcher, superset, subsetItem);
-            boolean optionalAndNotPresent = subsetItem.isOptional() && !containsKey(regexStringMatcher, superset, subsetItem);
-            if (!optionalAndNotPresent && subsetItemMatchingIndexes.isEmpty()) {
+            Set<Integer> subsetItemMatchingIndexes = matchesIndexes(regexStringMatcher, subsetItem, superset);
+            boolean optionalAndNotPresent = subsetItem.isOptional() && !containsKey(regexStringMatcher, subsetItem, superset);
+            boolean nottedAndPresent = nottedAndPresent(regexStringMatcher, subsetItem, superset);
+            if ((!optionalAndNotPresent && subsetItemMatchingIndexes.isEmpty()) || nottedAndPresent) {
                 result = false;
                 break;
             }
@@ -29,7 +33,7 @@ public class SubSetMatcher {
         return result;
     }
 
-    private static Set<Integer> matchesIndexes(RegexStringMatcher regexStringMatcher, List<ImmutableEntry> matchedList, ImmutableEntry matcherItem) {
+    private static Set<Integer> matchesIndexes(RegexStringMatcher regexStringMatcher, ImmutableEntry matcherItem, List<ImmutableEntry> matchedList) {
         Set<Integer> matchingIndexes = new HashSet<>();
         for (int i = 0; i < matchedList.size(); i++) {
             ImmutableEntry matchedItem = matchedList.get(i);
@@ -42,10 +46,24 @@ public class SubSetMatcher {
         return matchingIndexes;
     }
 
-    private static boolean containsKey(RegexStringMatcher regexStringMatcher, List<ImmutableEntry> matchedList, ImmutableEntry matcherItem) {
+    private static boolean containsKey(RegexStringMatcher regexStringMatcher, ImmutableEntry matcherItem, List<ImmutableEntry> matchedList) {
         for (ImmutableEntry matchedItem : matchedList) {
             if (regexStringMatcher.matches(matcherItem.getKey(), matchedItem.getKey(), true)) {
                 return true;
+            }
+        }
+        return false;
+    }
+
+    private static boolean nottedAndPresent(RegexStringMatcher regexStringMatcher, ImmutableEntry matcherItem, List<ImmutableEntry> matchedList) {
+        if (matcherItem.getKey().isNot()) {
+            NottableString unNottedMatcherItemKey = string(matcherItem.getKey().getValue());
+            for (ImmutableEntry matchedItem : matchedList) {
+                if (!matchedItem.getKey().isNot()) {
+                    if (regexStringMatcher.matches(unNottedMatcherItemKey, matchedItem.getKey(), true)) {
+                        return true;
+                    }
+                }
             }
         }
         return false;