+++ mockserver-core/src/test/java/org/mockserver/matchers/HeaderMatcherControlPlaneTest.java
@@ -107,6 +107,16 @@ public class HeaderMatcherControlPlaneTest {
         ), true).matches(
             null,
             new Headers().withEntries(
+                new Header("notHeaderOneName", "headerOneValue"),
+                new Header("notHeaderTwoName", "headerTwoValue")
+            )
+        ));
+        assertFalse(new MultiValueMapMatcher(new MockServerLogger(), new Headers().withEntries(
+            new Header(not("headerOneName"), "headerOneValue"),
+            new Header(not("headerTwoName"), not("headerTwoValue"))
+        ), true).matches(
+            null,
+            new Headers().withEntries(
                 new Header("headerOneName", "headerOneValue"),
                 new Header("headerTwoName", "headerTwoValue")
             )
@@ -309,7 +319,7 @@ public class HeaderMatcherControlPlaneTest {
             null,
             new Headers().withEntries(
                 new Header("headerOneName", "headerOneValueOne", "headerOneValueTwo"),
-                new Header("headerTwoName", "headerTwoValue")
+                new Header("notHeaderTwoName", "headerTwoValue")
             )
         ));
     }